requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
apscheduler==3.10.3
python-dotenv==1.0.0